[Java] Avoid int overflow when snapshots > 2GB. (#959)

+++ aeron-cluster/src/main/java/io/aeron/cluster/ClusterBackupAgent.java
@@ -538,7 +538,7 @@ public class ClusterBackupAgent implements Agent, UnavailableCounterHandler
         }
         else if (pollForResponse(clusterArchive, correlationId))
         {
-            final long snapshotStopPosition = (int)clusterArchive.controlResponsePoller().relevantId();
+            final long snapshotStopPosition = clusterArchive.controlResponsePoller().relevantId();
             correlationId = NULL_VALUE;
 
             if (NULL_POSITION == snapshotStopPosition)
